fix: Set default progress_callback to undefined
The 'pipeline' function from @xenova/transformers expects the 'progress_callback' option to be a function or undefined. The previous code used 'null' as the default value, which caused a type error during the build process.

This change updates the default value of 'progress_callback' to 'undefined' in the 'getInstance' method to align with the expected type and resolve the build failure.